pyproject.toml
[tool.poetry]
name = "tortoise-fastapi"
version = "0.0.1"
description = "The template for building scalable web APIs based on FastAPI, Tortoise ORM and others."
authors = ["prostomarkeloff"]
license = "MIT"
packages = [
    { include = "app" },
]


[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.111.0"
tortoise-orm = "^0.21.3"
pydantic = "^2.7.4"
asyncpg = "^0.29.0"
loguru = "^0.7.2"
betterconf = "^2.8.0"
# opentelemetry
opentelemetry-sdk = "^1.26.0"
opentelemetry-api = "^1.26.0"
opentelemetry-exporter-otlp-proto-grpc = "^1.26.0"
opentelemetry-exporter-otlp-proto-http = "^1.26.0"
opentelemetry-instrumentation-fastapi = "^0.47b0"
opentelemetry-instrumentation-tortoiseorm = "^0.47b0"

[tool.poetry.dev-dependencies]
black = {version = "^18.3-alpha.0", allows-prereleases = true}
pytest = "^5.3"
pytest-asyncio = "^0.10.0"
pre-commit = "^1.20"
pytest-cov = "^2.8"
mkdocs = "^1.0"
mkdocs-material = "^4.5"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
